+++ CHANGE_LOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add org.starchartlabs.alloy.core.Suppliers for streamlined and simplified use of Java Supplier construct
+- Add org.starchartlabs.alloy.core.Preconditions for streamlined validation of caller arguments